requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
Pillow==10.1.0
africastalking==1.2.0
web3==6.11.4
openai==1.3.7
speechrecognition==3.10.0
pyttsx3==2.90
requests==2.31.0
python-dotenv==1.0.0
cryptography==41.0.8
